requirements.txt
Django==3.2.4
phonenumberslite==8.12.29
django-crispy-forms==1.10.0
django-phonenumber-field==5.2.0
django-ckeditor-5==0.0.12